+++ code/engine.vc2008/xrGame/WeaponKnife.h
@@ -4,8 +4,7 @@
 #include "script_export_space.h"
 #include "../xrEngine/xr_collide_form.h"
 
-class CWeaponKnife: public CWeapon 
-{
+class CWeaponKnife: public CWeapon {
 private:
 	typedef CWeapon inherited;
 
@@ -26,7 +25,7 @@ protected:
 	virtual void		KnifeStrike					(const Fvector& pos, const Fvector& dir);
 
 	float				fWallmarkSize;
-	size_t				knife_material_idx;
+	u16					knife_material_idx;
 
 protected:
 	ALife::EHitType		m_eHitType;
