+++ code/engine.vc2008/xrGame/material_manager.h
@@ -1,7 +1,7 @@
 ////////////////////////////////////////////////////////////////////////////
 //	Module 		: material_manager.h
 //	Created 	: 27.12.2003
-//  Modified 	: 14.07.2017
+//  Modified 	: 27.12.2003
 //	Author		: Dmitriy Iassenev
 //	Description : Material manager
 ////////////////////////////////////////////////////////////////////////////
@@ -17,7 +17,7 @@ private:
 	bool						m_run_mode;
 	float						m_time_to_step;
 	u32							m_step_id;
-	size_t						m_my_material_idx;
+	u16							m_my_material_idx;
 	ref_sound					m_step_sound[4];
 	CObject						*m_object;
 	CPHMovementControl			*m_movement_control;
@@ -33,8 +33,8 @@ public:
 	virtual void				reload					(LPCSTR section);
 	virtual void				set_run_mode			(bool run_mode);
 	virtual void				update					(float time_delta, float volume, float step_time, bool standing);
-	IC		size_t				last_material_idx		() const;
-	IC		size_t				self_material_idx		() const;
+	IC		u16					last_material_idx		() const;
+	IC		u16					self_material_idx		() const;
 	IC		SGameMtlPair		*get_current_pair		();
 };
