+++ code/engine.vc2008/xrPhysics/ElevatorState.cpp
@@ -57,7 +57,7 @@ void CElevatorState::PhDataUpdate(float step)
 
 }
 
-void CElevatorState::InitContact(dContact* c, bool &do_collide, size_t , size_t)
+void CElevatorState::InitContact(dContact* c,bool &do_collide,u16 ,u16 )
 {
 
 }
@@ -390,7 +390,7 @@ bool CElevatorState::StateSwitchInertion(Estate new_state)
 	else return false;
 }
 
-bool	CElevatorState::UpdateMaterial(size_t &materil_idx)
+bool	CElevatorState::UpdateMaterial					( u16 &materil_idx )
 {
 	if( !ClimbingState() )
 		return false;
